Made dependency graph acyclic by removing the routes from the config.dart which was then renamed to init.dart to better reflect its use.
