+++ lib/menu/routes.dart
@@ -1,0 +1,6 @@
+class Routes {
+  static const String home = "/home";
+  static const String create = "/create";
+  static const String join = "/join";
+  static const String host = "/host";
+}
